% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractValueFromArrIndex}
\alias{extractValueFromArrIndex}
\title{Extracts the item at given index of an JSON array}
\usage{
extractValueFromArrIndex(arr, index)
}
\arguments{
\item{arr}{JSON array.}

\item{index}{Array index/key for which to retrieve value, zero-indexed.}
}
\value{
Character vector.
}
\description{
Extracts the item at given index of an JSON array
}
\keyword{internal}
