% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractValueFromObjKey.R
\name{extractValueFromObjKey}
\alias{extractValueFromObjKey}
\title{Extracts the item at given key of an JSON object}
\usage{
extractValueFromObjKey(obj, key)
}
\arguments{
\item{obj}{JSON object.}

\item{key}{Key for which to retrieve value.}
}
\value{
Character vector.
}
\description{
Extracts the item at given key of an JSON object
}
\keyword{internal}
