% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getKeysFromPaths}
\alias{getKeysFromPaths}
\title{Takes a split vector of dirty object keys and array indices and returns
 a 2d list providing references as required by \code{extractValueFrom*}.}
\usage{
getKeysFromPaths(paths)
}
\arguments{
\item{paths}{Character vector of jsonmatch-type key references.}
}
\value{
2d list providing symbols as required by \code{extractValueFrom*}.
}
\description{
Takes a split vector of dirty object keys and array indices and returns
 a 2d list providing references as required by \code{extractValueFrom*}.
}
\keyword{internal}

