% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getPathsFromPattern}
\alias{getPathsFromPattern}
\title{Splits pattern to paths while handling the wildcard}
\usage{
getPathsFromPattern(json, pattern)
}
\arguments{
\item{json}{JSON string.}

\item{pattern}{Subset pattern.}
}
\value{
Chr vector.
}
\description{
Splits pattern to paths while handling the wildcard
}
\keyword{internal}
