% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handleTrailingColon}
\alias{handleTrailingColon}
\title{Fixes a trailing colon in array selectors}
\usage{
handleTrailingColon(json, selector)
}
\arguments{
\item{json}{JSON string.}

\item{selector}{Object selector/path containing a trailing colon.}
}
\value{
Chr vector.
}
\description{
Fixes a trailing colon in array selectors
}
\keyword{internal}

