% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handleWildCard}
\alias{handleWildCard}
\title{Returns a chr vector of wildcard matched object keys}
\usage{
handleWildCard(json, selector)
}
\arguments{
\item{json}{JSON string.}

\item{selector}{Object selector/path containing the wildcard character.}
}
\value{
Chr vector.
}
\description{
Returns a chr vector of wildcard matched object keys
}
\keyword{internal}
