% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasUnclosedChar.R
\name{hasUnclosedChar}
\alias{hasUnclosedChar}
\title{Does a string contain a character neither enclosed in brackets nor
double quotes?}
\usage{
hasUnclosedChar(string, peep)
}
\arguments{
\item{string}{character. Vector of length 1.}

\item{peep}{character. Single character to search for. Cannot be any of
\code{[]{}\"}.}
}
\value{
Logical.
}
\description{
Does a string contain a character neither enclosed in brackets nor
double quotes?
}
\keyword{internal}
