% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isArray}
\alias{isArray}
\title{Is JSON a (non-empty) array(-like)?}
\usage{
isArray(json)
}
\arguments{
\item{json}{JSON string.}
}
\value{
Logical.
}
\description{
Is JSON a (non-empty) array(-like)?
}
\keyword{internal}
