% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isObject}
\alias{isObject}
\title{Is JSON a (non-empty) object(-like)?}
\usage{
isObject(json)
}
\arguments{
\item{json}{JSON string.}
}
\value{
Logical.
}
\description{
Is JSON a (non-empty) object(-like)?
}
\keyword{internal}
