% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTruthyChr}
\alias{isTruthyChr}
\title{Is character vector with number of characters >= 1?}
\usage{
isTruthyChr(x)
}
\arguments{
\item{x}{R object.}
}
\value{
Logical.
}
\description{
Is character vector with number of characters >= 1?
}
\keyword{internal}

