% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonmatch.R
\name{jsonmatch}
\alias{jsonmatch}
\title{Simple matching on JSON}
\usage{
jsonmatch(json, pattern, auto_unbox = FALSE, strict = TRUE)
}
\arguments{
\item{json}{Character. In-memory \code{JSON} string or filename.}

\item{pattern}{Character. Subset pattern.}

\item{auto_unbox}{Logical. Unbox \code{JSON}? Default: \code{FALSE}.}

\item{strict}{Logical. Only allow valid \code{JSON}? Default: \code{TRUE}.}
}
\value{
JSON string.
}
\description{
Subset JSON without parsing it to its equivalent R representation.
}
\details{
Parameter \code{pattern} allows matching keys of a JSON
object \code{.key,.k*y}, and keys/indices of a JSON array
\code{[0,3,5], [0:5], [0:]}.
...
}
