% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mutateInputJSON}
\alias{mutateInputJSON}
\title{Mutates input JSON for safe processing}
\usage{
mutateInputJSON(json)
}
\arguments{
\item{json}{Input JSON.}
}
\value{
JSON string.
}
\description{
Mutates input JSON for safe processing
}
\keyword{internal}

