% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{packStruct}
\alias{packStruct}
\title{Packs atomic data in arrays, optionally adding keys}
\usage{
packStruct(accu, json, paths)
}
\arguments{
\item{accu}{Character vector of target values (JSON extracts).}

\item{json}{Input JSON string.}

\item{paths}{Subset pattern paths.}
}
\value{
Character vector.
}
\description{
Packs atomic data in arrays, optionally adding keys
}
\keyword{internal}
