% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitOnUnclosedChar}
\alias{splitOnUnclosedChar}
\title{Splits a string on given character neither enclosed in brackets nor
double quotes}
\usage{
splitOnUnclosedChar(string, char, keep = FALSE)
}
\arguments{
\item{string}{Character vector of length 1L.}

\item{character}{Single character to split on.}
}
\value{
Chr vector.
}
\description{
Splits a string on given character neither enclosed in brackets nor
double quotes
}
\keyword{internal}

