% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stripArray}
\alias{stripArray}
\title{Strips an array's outer brackets}
\usage{
stripArray(json)
}
\arguments{
\item{json}{JSON array.}
}
\value{
Stripped JSON array.
}
\description{
Strips an array's outer brackets
}
\keyword{internal}
