% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stripObject}
\alias{stripObject}
\title{Strips an object's outer brackets}
\usage{
stripObject(json)
}
\arguments{
\item{json}{JSON object.}
}
\value{
Stripped JSON object.
}
\description{
Strips an object's outer brackets
}
\keyword{internal}

