% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{verifyPatternSyntax}
\alias{verifyPatternSyntax}
\title{Performs a syntax check on pattern}
\usage{
verifyPatternSyntax(json, pattern)
}
\arguments{
\item{json}{JSON string.}

\item{pattern}{Chr vector of length 1 specifying the subset pattern.}
}
\value{
Logical indicating whether the syntax is correct.
}
\description{
Performs a syntax check on pattern
}
\keyword{internal}

