% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasUnclosedChar}
\alias{hasUnclosedChar}
\title{Does a string contain a character neither enclosed in brackets nor
double quotes?}
\usage{
hasUnclosedChar(string, char)
}
\arguments{
\item{string}{Character vector of length 1L.}

\item{character}{Single character to search for.}
}
\value{
Logical.
}
\description{
Does a string contain a character neither enclosed in brackets nor
double quotes?
}
\keyword{internal}

