% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{isTruthyChrVec}
\alias{isTruthyChrVec}
\title{Is multi-length character vector with number of characters >= 1?}
\usage{
isTruthyChrVec(x)
}
\arguments{
\item{x}{R object.}
}
\value{
Logical.
}
\description{
Is multi-length character vector with number of characters >= 1?
}
\keyword{internal}
