% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitOnUnclosedChar.R
\name{splitOnUnclosedChar}
\alias{splitOnUnclosedChar}
\title{Splits a string on given character neither enclosed in brackets nor
double quotes}
\usage{
splitOnUnclosedChar(string, peep, keep = FALSE)
}
\arguments{
\item{string}{character. Vector of length 1.}

\item{peep}{character. Single character to split on. Cannot be any of
\code{[]{}"}.}

\item{keep}{logical. Keep \code{peep}?}
}
\value{
character. Vector of splits.
}
\description{
Splits a string on given character neither enclosed in brackets nor
double quotes
}
\keyword{internal}
